requirements.txt
kaggle>=1.5.16
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
seaborn>=0.12.0
onnx>=1.15.0
onnxruntime>=1.16.0

